+++ catscripts/Government/Countries/Australia/Australia.sh
@@ -1,20 +1,46 @@
 #!/bin/bash
 
-egrep -i 'Australia' newpages.txt | egrep -iv 'Canberra|Sydney|Melbourne' >> Australia.txt
+KEYWORDS_AUSTRALIA="Australia"
 
-AUSTRALIA=$(stat --print=%s Australia.txt)
-
-if [ $AUSTRALIA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Australia.txt"
-  export CATNAME="Australia"
-  $CATEGORIZE
+
+  debug_start "Australia"
+
+  AUSTRALIADIR="./catscripts/Government/Countries/Australia"
+
+  . $AUSTRALIADIR/Canberra/Canberra.sh #KEYWORDS_CANBERRA_ALL
+  . $AUSTRALIADIR/Sydney/Sydney.sh #KEYWORDS_SYDNEY_ALL
+  . $AUSTRALIADIR/Melbourne/Melbourne.sh #KEYWORDS_MELBOURNE_ALL
+
+  KEYWORDS_AUSTRALIA_EXCLUDE="$KEYWORDS_CANBERRA_ALL|$KEYWORDS_SYDNEY_ALL|$KEYWORDS_MELBOURNE_ALL"
+  KEYWORDS_AUSTRALIA_ALL="$KEYWORDS_AUSTRALIA|$KEYWORDS_AUSTRALIA_EXCLUDE"
+
+  AUSTRALIA=$(egrep -i "$KEYWORDS_AUSTRALIA" newpages.txt | egrep -iv "$KEYWORDS_AUSTRALIA_EXCLUDE")
+
+  if [ "$AUSTRALIA" != "" ];
+  then
+    printf "%s" "$AUSTRALIA" > Australia.txt
+    export CATFILE="Australia.txt"
+    export CATNAME="Australia"
+    $CATEGORIZE
+    rm Australia.txt
+    unset AUSTRALIA
+  fi
+
+  debug_end "Australia"
+
+else
+
+    AUSTRALIADIR="./catscripts/Government/Countries/Australia"
+
+  . $AUSTRALIADIR/Canberra/Canberra.sh norun #KEYWORDS_CANBERRA_ALL
+  . $AUSTRALIADIR/Sydney/Sydney.sh norun #KEYWORDS_SYDNEY_ALL
+  . $AUSTRALIADIR/Melbourne/Melbourne.sh norun #KEYWORDS_MELBOURNE_ALL
+
+  KEYWORDS_AUSTRALIA_EXCLUDE="$KEYWORDS_CANBERRA_ALL|$KEYWORDS_SYDNEY_ALL|$KEYWORDS_MELBOURNE_ALL"
+  KEYWORDS_AUSTRALIA_ALL="$KEYWORDS_AUSTRALIA|$KEYWORDS_AUSTRALIA_EXCLUDE"
+
 fi
 
-CURRENTDIR="./catscripts/Government/Countries/Australia"
-
-$CURRENTDIR/Canberra/Canberra.sh
-$CURRENTDIR/Sydney/Sydney.sh
-$CURRENTDIR/Melbourne/Melbourne.sh
-
-rm Australia.txt
+
