+++ catscripts/Government/Countries/Australia/Sydney/Sydney.sh
@@ -1,24 +1,38 @@
 #!/bin/bash
 
-egrep -i 'Sydney(,|)(| )Australia|Sydney Harbo(u|)r|Australia.+Sydney|Sydney.+Australia' newpages.txt | egrep -iv 'Sydney Opera House' >> Sydney.txt
-egrep -i 'Sydney(| )Opera(| )House' newpages.txt >> SydneyOperaHouse.txt
+KEYWORDS_SYDNEY="Sydney(,|)(| )Australia|Sydney Harbo(u|)r|Australia.+Sydney|Sydney.+Australia"
+KEYWORDS_OPERAHOUSE="Sydney(| )Opera(| )House"
+KEYWORDS_SYDNEY_EXCLUDE="$KEYWORDS_OPERAHOUSE"
+KEYWORDS_SYDNEY_ALL="$KEYWORDS_SYDNEY|$KEYWORDS_OPERAHOUSE"
 
-SYDNEY=$(stat --print=%s Sydney.txt)
-OPERA=$(stat --print=%s SydneyOperaHouse.txt)
-
-if [ $SYDNEY -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Sydney.txt"
-  export CATNAME="Sydney, Australia"
-  $CATEGORIZE
-fi
 
-if [ $OPERA -ne 0 ];
-then
-  export CATFILE="SydneyOperaHouse.txt"
-  export CATNAME="Sydney Opera House"
-  $CATEGORIZE
-fi
+  debug_start "Sydney, Australia"
 
-rm SydneyOperaHouse.txt
-rm Sydney.txt
+  SYDNEY=$(egrep -i "$KEYWORDS_SYDNEY" newpages.txt | egrep -iv "$KEYWORDS_SYDNEY_EXCLUDE")
+  OPERA=$(egrep -i "$KEYWORDS_OPERAHOUSE" newpages.txt)
+
+  if [ "$SYDNEY" != "" ];
+  then
+    printf "%s" "$SYDNEY" > Sydney.txt
+    export CATFILE="Sydney.txt"
+    export CATNAME="Sydney, Australia"
+    $CATEGORIZE
+    rm Sydney.txt
+    unset SYDNEY
+  fi
+
+  if [ "$OPERA" != "" ];
+  then
+    printf "%s" "$OPERA" > SydneyOperaHouse.txt
+    export CATFILE="SydneyOperaHouse.txt"
+    export CATNAME="Sydney Opera House"
+    $CATEGORIZE
+    rm SydneyOperaHouse.txt
+    unset OPERA
+  fi
+
+  debug_end "Sydney, Australia"
+
+fi
