+++ catscripts/Government/Countries/Bangladesh/Bangladesh.sh
@@ -1,14 +1,25 @@
 #!/bin/bash
 
-egrep -i 'Bangladesh' newpages.txt >> Bangladesh.txt
+KEYWORDS_BANGLADESH="Bangladesh"
+KEYWORDS_BANGLADESH_ALL="$KEYWORDS_BANGLADESH"
 
-BANGLADESH=$(stat --print=%s Bangladesh.txt)
-
-if [ $BANGLADESH -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Bangladesh.txt"
-  export CATNAME="Bangladesh"
-  $CATEGORIZE
-fi
 
-rm Bangladesh.txt
+  debug_start "Bangladesh"
+
+  BANGLADESH=$(egrep -i "$KEYWORDS_BANGLADESH" newpages.txt)
+
+  if [ "$BANGLADESH" != "" ];
+  then
+    printf "%s" "$BANGLADESH" > Bangladesh.txt
+    export CATFILE="Bangladesh.txt"
+    export CATNAME="Bangladesh"
+    $CATEGORIZE
+    rm Bangladesh.txt
+    unset BANGLADESH
+  fi
+
+  debug_end "Bangladesh"
+
+fi
