+++ catscripts/Government/Countries/United_Arab_Emirates/UAE.sh
@@ -1,24 +1,38 @@
 #!/bin/bash
 
-egrep -i '\bUAE|\bU\.A\.E|United(| )Arab(| )Emirates' newpages.txt | egrep -iv 'Dubai' >> UAE.txt
-egrep -i 'Dubai' newpages.txt >> Dubai.txt
+KEYWORDS_UAE="\bUAE|\bU\.A\.E|United(| )Arab(| )Emirates"
+KEYWORDS_DUBAI="Dubai"
+KEYWORDS_UAE_EXCLUDE="$KEYWORDS_DUBAI"
+KEYWORDS_UAE_ALL="$KEYWORDS_UAE|$KEYWORDS_DUBAI"
 
-UAE=$(stat --print=%s UAE.txt)
-DUBAI=$(stat --print=%s Dubai.txt)
-
-if [ $UAE -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="UAE.txt"
-  export CATNAME="United Arab Emirates"
-  $CATEGORIZE
-fi
+  
+  debug_start "United Arab Emirates"
 
-if [ $DUBAI -ne 0 ];
-then
-  export CATFILE="Dubai.txt"
-  export CATNAME="Dubai"
-  $CATEGORIZE
-fi
+  UAE=$(egrep -i "$KEYWORDS_UAE" newpages.txt | egrep -iv "$KEYWORDS_UAE_EXCLUDE")
+  DUBAI=$(egrep -i "$KEYWORDS_DUBAI" newpages.txt)
 
-rm UAE.txt
-rm Dubai.txt
+  if [ "$UAE" != "" ];
+  then
+    printf "%s" "$UAE" > UAE.txt
+    export CATFILE="UAE.txt"
+    export CATNAME="United Arab Emirates"
+    $CATEGORIZE
+    rm UAE.txt
+    unset UAE
+  fi
+
+  if [ "$DUBAI" != "" ];
+  then
+    printf "%s" "$DUBAI" > Dubai.txt
+    export CATFILE="Dubai.txt"
+    export CATNAME="Dubai"
+    $CATEGORIZE
+    rm Dubai.txt
+    unset DUBAI
+  fi
+
+  debug_end "United Arab Emirates"
+
+fi
